+++ src/java/quickfix_MySQLLog.cpp
@@ -44,7 +44,7 @@ JNIEXPORT void JNICALL Java_quickfix_MySQLLog_destroy
 
   JVM::set( pEnv );
   JVMObject jobject( obj );
-  FIX::Log* pLog = ( FIX::Log* ) jobject.getInt( "cppPointer" );
+  FIX::Log* pLog = ( FIX::Log* ) jobject.getLong( "cppPointer" );
   delete pLog;
 
   QF_STACK_CATCH
