+++ src/java/quickfix_ScreenLog.cpp
@@ -41,7 +41,7 @@ JNIEXPORT void JNICALL Java_quickfix_ScreenLog_destroy
 
   JVM::set( pEnv );
   JVMObject jobject( obj );
-  FIX::Log* pLog = ( FIX::Log* ) jobject.getInt( "cppPointer" );
+  FIX::Log* pLog = ( FIX::Log* ) jobject.getLong( "cppPointer" );
   delete pLog;
 
   QF_STACK_CATCH
